Fix windows linking issue by having separate visibility macro
