+++ test_msgs/include/test_msgs/action/fibonacci__type_support.h
@@ -25,11 +25,11 @@ extern "C"
 #include "rosidl_generator_c/action_type_support_struct.h"
 #include "rosidl_typesupport_interface/macros.h"
 
-#include "test_msgs/msg/rosidl_generator_c__visibility_control.h"
+#include "test_msgs/action/rosidl_generator_c__visibility_control.h"
 
 /* *INDENT-OFF* */
 // Forward declare the get type support functions for this type.
-ROSIDL_GENERATOR_C_PUBLIC_test_msgs
+ROSIDL_GENERATOR_C_PUBLIC_test_msgs_ACTION
 const rosidl_action_type_support_t *
 ROSIDL_TYPESUPPORT_INTERFACE__ACTION_SYMBOL_NAME(
   rosidl_typesupport_c, test_msgs, action, Fibonacci)();
